chore: moved contactModal in modals folder
